+++ .gitignore
@@ -1,3 +1,4 @@
 _site
 .DS_Store
 node_modules
+.jekyll_cache/
