+++ jekyll/_layouts/default.html
@@ -32,7 +32,7 @@
 	              <a class="nav-link" href="/">Home</span></a>
 	            </li>
 	            <li class="nav-item">
-	              <a class="nav-link" href="/dnd-exchange-rate">Coin Exchange</a>
+	              <a class="nav-link" href="/dnd-exchange-rate">Coin Coverter Calculator</a>
 	            </li>
 	            <li class="nav-item">
 	              <a class="nav-link" href="/dnd-travel-calculator">Travel Calculator</a>
