+++ react/components/calculator/coin/PerPersonResult.js
